libwasi: move a few things into wasi_subr.[ch]
