+++ libwasi/wasi_host_subr.c
@@ -97,26 +97,6 @@ wasi_convert_filestat(const struct stat *hst, struct wasi_filestat *wst)
 #endif
 }
 
-int
-wasi_unstable_convert_filestat(const struct wasi_filestat *wst,
-                               struct wasi_unstable_filestat *uwst)
-{
-        uint64_t linkcount = le64_decode(&wst->linkcount);
-        if (linkcount > UINT32_MAX) {
-                return E2BIG;
-        }
-        memset(uwst, 0, sizeof(*uwst));
-        uwst->dev = wst->dev;
-        uwst->ino = wst->ino;
-        uwst->type = wst->type;
-        le32_encode(&uwst->linkcount, linkcount);
-        uwst->size = wst->size;
-        uwst->atim = wst->atim;
-        uwst->mtim = wst->mtim;
-        uwst->ctim = wst->ctim;
-        return 0;
-}
-
 uint8_t
 wasi_convert_dirent_filetype(uint8_t hosttype)
 {
